+++ app/src/main/java/s/yarlykov/izisandbox/recycler_and_swipes/time_line_2d/ZoomConsumer.kt
@@ -1,5 +1,5 @@
 package s.yarlykov.izisandbox.recycler_and_swipes.time_line_2d
 
 interface ZoomConsumer {
-    fun onZoomChanged(zoom : Int)
+    fun onZoomChanged(zoom : Float)
 }
